+++ mkl_fft/interfaces/_scipy_fft.py
@@ -39,7 +39,7 @@ import numpy as np
 
 import mkl_fft
 
-from .._fft_utils import _compute_fwd_scale, _swap_direction
+from .._fft_utils import _swap_direction
 from ._float_utils import _supported_array_or_not_implemented
 
 __all__ = [
@@ -233,10 +233,9 @@ def fft(
     _check_plan(plan)
     x = _validate_input(x)
     out = _use_input_as_out(x, overwrite_x)
-    fsc = _compute_fwd_scale(norm, n, x.shape[axis])
 
     with _Workers(workers):
-        return mkl_fft.fft(x, n=n, axis=axis, fwd_scale=fsc, out=out)
+        return mkl_fft.fft(x, n=n, axis=axis, norm=norm, out=out)
 
 
 def ifft(
@@ -251,10 +250,9 @@ def ifft(
     _check_plan(plan)
     x = _validate_input(x)
     out = _use_input_as_out(x, overwrite_x)
-    fsc = _compute_fwd_scale(norm, n, x.shape[axis])
 
     with _Workers(workers):
-        return mkl_fft.ifft(x, n=n, axis=axis, fwd_scale=fsc, out=out)
+        return mkl_fft.ifft(x, n=n, axis=axis, norm=norm, out=out)
 
 
 def fft2(
@@ -331,10 +329,9 @@ def fftn(
     x = _validate_input(x)
     out = _use_input_as_out(x, overwrite_x)
     s, axes = _init_nd_shape_and_axes(x, s, axes)
-    fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
-        return mkl_fft.fftn(x, s=s, axes=axes, fwd_scale=fsc, out=out)
+        return mkl_fft.fftn(x, s=s, axes=axes, norm=norm, out=out)
 
 
 def ifftn(
@@ -357,10 +354,9 @@ def ifftn(
     x = _validate_input(x)
     out = _use_input_as_out(x, overwrite_x)
     s, axes = _init_nd_shape_and_axes(x, s, axes)
-    fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
-        return mkl_fft.ifftn(x, s=s, axes=axes, fwd_scale=fsc, out=out)
+        return mkl_fft.ifftn(x, s=s, axes=axes, norm=norm, out=out)
 
 
 def rfft(
@@ -374,11 +370,10 @@ def rfft(
     """
     _check_plan(plan)
     x = _validate_input(x)
-    fsc = _compute_fwd_scale(norm, n, x.shape[axis])
 
     with _Workers(workers):
         # Note: overwrite_x is not utilized
-        return mkl_fft.rfft(x, n=n, axis=axis, fwd_scale=fsc)
+        return mkl_fft.rfft(x, n=n, axis=axis, norm=norm)
 
 
 def irfft(
@@ -392,11 +387,10 @@ def irfft(
     """
     _check_plan(plan)
     x = _validate_input(x)
-    fsc = _compute_fwd_scale(norm, n, 2 * (x.shape[axis] - 1))
 
     with _Workers(workers):
         # Note: overwrite_x is not utilized
-        return mkl_fft.irfft(x, n=n, axis=axis, fwd_scale=fsc)
+        return mkl_fft.irfft(x, n=n, axis=axis, norm=norm)
 
 
 def rfft2(
@@ -472,11 +466,10 @@ def rfftn(
     _check_plan(plan)
     x = _validate_input(x)
     s, axes = _init_nd_shape_and_axes(x, s, axes)
-    fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
         # Note: overwrite_x is not utilized
-        return mkl_fft.rfftn(x, s, axes, fwd_scale=fsc)
+        return mkl_fft.rfftn(x, s, axes, norm=norm)
 
 
 def irfftn(
@@ -498,11 +491,10 @@ def irfftn(
     _check_plan(plan)
     x = _validate_input(x)
     s, axes = _init_nd_shape_and_axes(x, s, axes, invreal=True)
-    fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
         # Note: overwrite_x is not utilized
-        return mkl_fft.irfftn(x, s, axes, fwd_scale=fsc)
+        return mkl_fft.irfftn(x, s, axes, norm=norm)
 
 
 def hfft(
@@ -520,11 +512,10 @@ def hfft(
     norm = _swap_direction(norm)
     x = np.array(x, copy=True)
     np.conjugate(x, out=x)
-    fsc = _compute_fwd_scale(norm, n, 2 * (x.shape[axis] - 1))
 
     with _Workers(workers):
         # Note: overwrite_x is not utilized
-        return mkl_fft.irfft(x, n=n, axis=axis, fwd_scale=fsc)
+        return mkl_fft.irfft(x, n=n, axis=axis, norm=norm)
 
 
 def ihfft(
@@ -539,11 +530,10 @@ def ihfft(
     _check_plan(plan)
     x = _validate_input(x)
     norm = _swap_direction(norm)
-    fsc = _compute_fwd_scale(norm, n, x.shape[axis])
 
     with _Workers(workers):
         # Note: overwrite_x is not utilized
-        result = mkl_fft.rfft(x, n=n, axis=axis, fwd_scale=fsc)
+        result = mkl_fft.rfft(x, n=n, axis=axis, norm=norm)
 
     np.conjugate(result, out=result)
     return result
@@ -626,11 +616,10 @@ def hfftn(
     x = np.array(x, copy=True)
     np.conjugate(x, out=x)
     s, axes = _init_nd_shape_and_axes(x, s, axes, invreal=True)
-    fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
         # Note: overwrite_x is not utilized
-        return mkl_fft.irfftn(x, s, axes, fwd_scale=fsc)
+        return mkl_fft.irfftn(x, s, axes, norm=norm)
 
 
 def ihfftn(
@@ -653,11 +642,10 @@ def ihfftn(
     x = _validate_input(x)
     norm = _swap_direction(norm)
     s, axes = _init_nd_shape_and_axes(x, s, axes)
-    fsc = _compute_fwd_scale(norm, s, x.shape)
 
     with _Workers(workers):
         # Note: overwrite_x is not utilized
-        result = mkl_fft.rfftn(x, s, axes, fwd_scale=fsc)
+        result = mkl_fft.rfftn(x, s, axes, norm=norm)
 
     np.conjugate(result, out=result)
     return result